back/requirements.txt
Flask
firebase-admin
google-generativeai
Flask-Cors
itsdangerous

